Add main navigation to the project
Because:
- Main navigation needed for full site navigation
- Navigation markup needed to guide styling
- Structure of site theme toggle button needed

I did the primary navigation required for this project along with a
theme toggle button.

NOTE - Previous commit on feature branch failed because my laptop
powered off in the middle of the commit process as a result of low
battery.

NOTE - DIDN'T PROPERLY ADD NEW FILES IN PREVIOUS COMMITS. CORRECTED THE
ERROR IN THIS COMMIT.