Setup this project to use Eslint & Prettier
Because:
- Linting is necessary to prevent bugs
- It enforces established best practices
- Prettier affords consistent code styling

Eslint and Prettier packages play well in affording a project a well-
defined function and form benchmark right from the get go. Having both
in this project means I can code rest assured that I am following the
the guidelines some of the smartest people on earth have defined.