Add context for the home/about combo
Because:
- I need only a portion of info on the homepage to change
- I want the about page content to have my profile as image
- I need to manage which one displays through state

In this commit I introduce a Home context to the project. This is
where I manage the state that determines which one between the
homepage content and the about page content that slids up beside my
profile image on the home page.

I needed to do this so the nav component can access that context and
dynamically update the state handling the content display based on
which nav link between Home and About was clicked.