Flesh out contact page
Because:
- Former state was skeletal
- I needed to flesh it out so it looks functional

I built out the contact page with a form and a series of
social media anchors that link to my social media.

What I have is skeletal and needsmore work.

I'll need to implement the form elements the react way.
And I'll also need to refactor the hard style px values
to use responsive and adaptive units, miuch like the rest
of the project.