Add context to portfolio page
Because:
- I need a simple way to loop through my porfolio
- I may raise and use and api for it in future

I didn't want to manually code all the projects I wish to
include in this site. Moreover, the projects on display here
will grow over time. Using an object literal to add new
projects makes more sense to me.

Also, I may use an api in the future to fetch the projects
I want to include. So using a distinct context for that
is seems smart to me.